For testing/experimental reduce the number of good_signatures
We want to increase the speed of experimental/testing releases. To
reduce the time necessary to sign each image by two separate persons,
reduce the required number of good signatures to 1.

This does not influence the stable releases. These will still require 2
signatures to be confident about release quality and in case of leaked
keys.

Signed-off-by: Markus Pargmann <mpa@devmp.org>